Mejoras en comentarios

+++ controlador/ControladorConsultaPedido.php
@@ -1,4 +1,16 @@
 <?php
+/**
+ * Controlador para el manejo de la pantalla de consulta de pedido y carrito
+ *
+ * Clase php que hereda de ControladorBase añadiendo funcionalidad concreta para el manejo de la consulta de pedidos y carrito de la compra
+ * Recibe los eventos que se producen en la vista de consulta de pedido (carrito), los gestiona y realiza las peticiones correspondientes 
+ * a los modelos de producto, item, pedido e items del pedido.
+ * 
+ * @author Gerard Herrera Sague
+ * @author Paul Morrison Aguiar
+ * @author Jesús Pérez Melero
+ * @author Carmen María Candal alonso
+*/
 // incluimos los modelos a utilizar
 require_once('modelo/ModeloPedido.php');
 require_once('modelo/ModeloItem.php');
@@ -23,6 +35,9 @@ class ControladorConsultaPedido extends ControladorBase
 
     }
 
+    //Obtiene la información de un pedido
+    // idPedido: identificador del pedido 
+    // Se carga la vista de consulta del pedido.
     public function obtenerPedido($idPedido)
     {
 
@@ -52,50 +67,57 @@ class ControladorConsultaPedido extends ControladorBase
         $this->view("VistaConsultaPedido", array());
     }
 
+    //Obtiene la lista de pedidos realizados por ese usuario 
+    // Se carga la vista de consulta de pedido con esa información
     public function obtenerPedidos() {
 
-        unset($_SESSION['pedidoActual']);
-        unset($_SESSION['idPedidoActual']);
+        if (isset($_SESSION['pedidoActual'])) unset($_SESSION['pedidoActual']);
+        if (isset($_SESSION['idPedidoActual'])) unset($_SESSION['idPedidoActual']);
         $this->cargarDatosPedidos();
         $this->view("VistaConsultaPedido", array());
     }
 
+    //Carga los datos de los pedidos de un usuario
     private function cargarDatosPedidos() {
 
-        $imgs = array();
+        $imgs = array();        
         $idUsuario = $_SESSION['userId'];
-        $idPedidoCarrito = $_SESSION['idPedido'];
+        if (isset($_SESSION['idPedido'])){
+            $idPedidoCarrito = $_SESSION['idPedido'];
 
-        //Obtenemos los pedidos
-        $pedidosCompletados = $this->pedidoRepository->getPedidosCompletados($idUsuario);
+            //Obtenemos los pedidos
+            $pedidosCompletados = $this->pedidoRepository->getPedidosCompletados($idUsuario);
 
-        if(isset($pedidosCompletados) && (!$pedidosCompletados=="")) {
-            $_SESSION['pedidosCompletados'] = $pedidosCompletados;
-        } else {
-            $_SESSION['pedidosCompletados'] = null;
-        }
-
-        if(isset($idPedidoCarrito)) {
-
-            $carrito = $this->pedidoRepository->readById($idPedidoCarrito);
-
-            //Obtenemos los items del carrito
-            $itemsCarrito = array();
-            $i = 0;
-            $itemsPedido = $this->itemsPedidoRepository->readBy("id_pedido", $carrito->getId());
-            foreach($itemsPedido as $itemPedido) {
-
-                $item = $this->itemRepository->readById($itemPedido->getId_item());
-                $product = $this->product_modelo->readById($item->getId_categoria());
-                $imgs[$product->getId()] = $product->getImg();
-                $itemsCarrito[$i++] = $item;
-
+            if(isset($pedidosCompletados) && (!$pedidosCompletados=="")) {
+                $_SESSION['pedidosCompletados'] = $pedidosCompletados;
+            } else {
+                $_SESSION['pedidosCompletados'] = null;
             }
-            $_SESSION['carrito'] = $itemsCarrito;
-            $_SESSION['imgs'] = $imgs;
+
+            if(isset($idPedidoCarrito)) {
+
+                $carrito = $this->pedidoRepository->readById($idPedidoCarrito);
+
+                //Obtenemos los items del carrito
+                $itemsCarrito = array();
+                $i = 0;
+                $itemsPedido = $this->itemsPedidoRepository->readBy("id_pedido", $carrito->getId());
+                foreach($itemsPedido as $itemPedido) {
+
+                    $item = $this->itemRepository->readById($itemPedido->getId_item());
+                    $product = $this->product_modelo->readById($item->getId_categoria());
+                    $imgs[$product->getId()] = $product->getImg();
+                    $itemsCarrito[$i++] = $item;
+
+                }
+                $_SESSION['carrito'] = $itemsCarrito;
+                $_SESSION['imgs'] = $imgs;
+            }
         }
     }
 
+    //Elimina un item del pedido
+    // idItem: identificador único de un item de un pedido
     public function remove($idItem)
     {
         if (!isset($idItem)) return;
@@ -117,33 +139,39 @@ class ControladorConsultaPedido extends ControladorBase
         $this->obtenerPedidos();
     }
 
+    // Confirma el pedido 
+    // Modifica el estado del pedido a "completado", limpia los elementos del carrito y recarga la lista de pedidos
     public function confirmarPedido() {
-
-        $idPedidoCarrito = $_SESSION['idPedido'];
-        $pedido = $this->pedidoRepository->readById($idPedidoCarrito);
-        $pedido->setEstado("Completado");
-        $this->pedidoRepository->updatePedido($pedido);
-        $this->limpiarCarrito();
-        $this->obtenerPedidos();
+        if (isset($_SESSION['idPedido'])){
+            $idPedidoCarrito = $_SESSION['idPedido'];
+            $pedido = $this->pedidoRepository->readById($idPedidoCarrito);
+            $pedido->setEstado("Completado");
+            $this->pedidoRepository->updatePedido($pedido);
+            $this->limpiarCarrito();
+            $this->obtenerPedidos();
+        }
 
     }
 
+    // Vacía los items del carrito
     public function vaciarCarrito() {
 
         $this->limpiarCarrito();
         $this->obtenerPedidos();
-
-        $idPedidoCarrito = $_SESSION['idPedido'];
-        $this->pedidoRepository->deleteById($idPedidoCarrito);
+        if (isset($_SESSION['idPedido'])){
+            $idPedidoCarrito = $_SESSION['idPedido'];
+            $this->pedidoRepository->deleteById($idPedidoCarrito);
+        }
     }
 
+    // Elimina de la sesión la información del pedido actual
     private function limpiarCarrito() {
-        unset($_SESSION['idPedido']);
-        unset($_SESSION['numItemsPedido']);
-        unset($_SESSION['totalPedido']);
-        unset($_SESSION['carrito']);
-        unset($_SESSION['pedidoActual']);
-        unset($_SESSION['idPedidoActual']);
+        if (isset($_SESSION['idPedido'])) unset($_SESSION['idPedido']);
+        if (isset($_SESSION['numItemsPedido'])) unset($_SESSION['numItemsPedido']);
+        if (isset($_SESSION['totalPedido'])) unset($_SESSION['totalPedido']);
+        if (isset($_SESSION['carrito'])) unset($_SESSION['carrito']);
+        if (isset($_SESSION['pedidoActual'])) unset($_SESSION['pedidoActual']);
+        if (isset($_SESSION['idPedidoActual'])) unset($_SESSION['idPedidoActual']);
     }
 }
